Move expiry date to a new package
