+++ src/main/java/seedu/storemando/logic/parser/AddCommandParser.java
@@ -6,14 +6,14 @@ import static seedu.storemando.logic.parser.CliSyntax.PREFIX_LOCATION;
 import static seedu.storemando.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.storemando.logic.parser.CliSyntax.PREFIX_QUANTITY;
 import static seedu.storemando.logic.parser.CliSyntax.PREFIX_TAG;
-import static seedu.storemando.model.item.ExpiryDate.NO_EXPIRY_DATE;
+import static seedu.storemando.model.expirydate.ExpiryDate.NO_EXPIRY_DATE;
 
 import java.util.Set;
 import java.util.stream.Stream;
 
 import seedu.storemando.logic.commands.AddCommand;
 import seedu.storemando.logic.parser.exceptions.ParseException;
-import seedu.storemando.model.item.ExpiryDate;
+import seedu.storemando.model.expirydate.ExpiryDate;
 import seedu.storemando.model.item.Item;
 import seedu.storemando.model.item.ItemName;
 import seedu.storemando.model.item.Location;
