+++ src/main/java/seedu/storemando/model/expirydate/ExpiryDate.java
@@ -1,4 +1,4 @@
-package seedu.storemando.model.item;
+package seedu.storemando.model.expirydate;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.storemando.commons.util.AppUtil.checkArgument;
