+++ src/main/java/seedu/storemando/model/item/Item.java
@@ -7,6 +7,7 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+import seedu.storemando.model.expirydate.ExpiryDate;
 import seedu.storemando.model.tag.Tag;
 
 /**
