+++ src/test/java/seedu/storemando/logic/commands/AddCommandTest.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.storemando.model.item.ExpiryDate.NO_EXPIRY_DATE;
+import static seedu.storemando.model.expirydate.ExpiryDate.NO_EXPIRY_DATE;
 import static seedu.storemando.testutil.Assert.assertThrows;
 
 import java.nio.file.Path;
