+++ src/test/java/seedu/storemando/logic/parser/EditCommandParserTest.java
@@ -36,7 +36,7 @@ import org.junit.jupiter.api.Test;
 import seedu.storemando.commons.core.index.Index;
 import seedu.storemando.logic.commands.EditCommand;
 import seedu.storemando.logic.commands.EditCommand.EditItemDescriptor;
-import seedu.storemando.model.item.ExpiryDate;
+import seedu.storemando.model.expirydate.ExpiryDate;
 import seedu.storemando.model.item.ItemName;
 import seedu.storemando.model.item.Location;
 import seedu.storemando.model.item.Quantity;
